opt:all name
The name of the previous version of the file was misnamed, and I made a new change to all the files.

There was a problem with the file hierarchy in the previous version, and I made new changes.

In addition, I would like to submit an assignment on JS references.